Add link to error message to continue user journey
The additional link provides the ability to continue without beeing lost while he already subscribed to a plan, although he doesnt know or remembers. This way the user gets the ability to continue to the course or even go back and chose a different course as he now realised he picked the wrong one. A redirect in order to push the user towards the course would be also possible but does not provide the chance to let the user decide what he really wants. Looking forward to your thoughts about this request.