Update privacy policy (#1021)
* Updated exporters to include Group Descriptions

Adds support for group_description for privacy exporters which was added in WP5.3 through WPCoreChangeset#45825 and WPCoreTracTicket#45491

* Update Privacy Policy Guide content to drop use of deprecated classes

Update Suggested Privacy Policy text to utilize privacy-policy-tutorial css class instead of wp-policy-help as it was deprecated. Also wrap the contents in the wp-suggested-text div to style the section to match WordPress. This follows from [WPCoreTrac#49282](https://core.trac.wordpress.org/ticket/49282) and although back-compat is being introduced in WP5.4 as of [WPChangeset#47112](https://core.trac.wordpress.org/changeset/47112) this change will better support users of WP5.1-5.4

This also removes the wrapping <div contenteditable="false"> as it's now unnecessary. As long as the content uses the proper privacy-policy-tutorial class then the WP Privacy Policy Guide section copy action will avoid copying that content. The contenteditable divs were there originally when the contents was fully copied due to not using the appropriate class, and when originally the entire guide would be dumped into the WYSIWYG editor with wp-policy-help sections being highlighted.

* Replace preceding spaces with tabs

* Fix plural typo

* Added new @since to indicate what's updated

Applied `* @since    3.37.8 Added `$group_description` to the group exporter.` to all updated exporters.

* Added updated @since

* update docs

* fix cs

* strict compare page numbers

* fix typo

Co-authored-by: Garrett Hyder <garrett.m.hyder@gmail.com>