Don't merge open PRs in dev by default.
We're not sure why we would want to merge open PRs in the dev branch AFTER pushing a release. This commit disables that by setting the default for the 'merge' option to ''.

This didn't seem to be an issue in the past because failing tests and branch protection rules would stop any PRs from merging anyway.